terraform-provider-sdm: correctly handle required and int32 fields
#4971

GitOrigin-RevId: 344b0f7a96eb053d9a078ffd853f4758605aa1d1

+++ helpers.go
@@ -27,7 +27,7 @@ func int32FromMap(m map[string]interface{}, key string) int32 {
 	if value == nil {
 		return 0
 	}
-	return value.(int32)
+	return int32(value.(int))
 }
 
 func boolFromMap(m map[string]interface{}, key string) bool {
@@ -51,7 +51,7 @@ func int32FromResourceData(d *schema.ResourceData, key string) int32 {
 	if value == nil {
 		return 0
 	}
-	return value.(int32)
+	return int32(value.(int))
 }
 
 func boolFromResourceData(d *schema.ResourceData, key string) bool {
